Use config.json, move main.py & definitions.py
