+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__/
 .vscode/
 *.env
 venv/
+config.json
