+++ config.example.json
@@ -1,0 +1,4 @@
+{
+    "render_dir": "D:/TrainingGrounds-Graphics/Character Renders",
+    "output_dir": "C:/Users/your_name/Desktop/"
+}
